scripts/train_model.py
import os
import pandas as pd
import joblib # For saving the model
from sklearn.linear_model import LogisticRegression # Example model
from sklearn.metrics import accuracy_score
import argparse

def train_model_main(processed_data_path: str, trained_model_path: str, training_metrics_path: str):
    """
    Simulates model training and saves the trained model and metrics.

    Args:
        processed_data_path: The input path for the processed data (directory).
        trained_model_path: The output path for the trained model (directory).
        training_metrics_path: The output path for the training metrics (file).
    """
    print(f"Simulating model training using data from {processed_data_path}...")

    # Load dummy processed data (assuming it's a CSV inside the directory)
    processed_data_file = os.path.join(processed_data_path, "processed_data.csv")
    if not os.path.exists(processed_data_file):
        raise FileNotFoundError(f"Processed data file not found: {processed_data_file}")
    
    data = pd.read_csv(processed_data_file)
    
    # Prepare features (X) and target (y)
    X = data[["feature_a", "feature_b"]]
    y = data["target"]

    # Simulate training a simple model
    model = LogisticRegression(random_state=42)
    model.fit(X, y)
    
    predictions = model.predict(X)
    accuracy = accuracy_score(y, predictions)

    print(f"Model trained. Simulated accuracy: {accuracy:.4f}")

    # Ensure output directories exist
    os.makedirs(trained_model_path, exist_ok=True)
    os.makedirs(os.path.dirname(training_metrics_path), exist_ok=True)

    # Save the trained model
    model_file = os.path.join(trained_model_path, "model.joblib")
    joblib.dump(model, model_file)
    print(f"Trained model saved to: {model_file}")

    # Save training metrics
    with open(training_metrics_path, "w") as f:
        f.write("Training Metrics Summary:\n")
        f.write(f"  Model Type: Logistic Regression\n")
        f.write(f"  Simulated Accuracy: {accuracy:.4f}\n")
    print(f"Training metrics report saved to: {training_metrics_path}")

if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Train model component.")
    parser.add_argument("--processed-data-path", type=str, required=True,
                        help="Path to the processed data directory.")
    parser.add_argument("--trained-model-path", type=str, required=True,
                        help="Output path for the trained model directory.")
    parser.add_argument("--training-metrics-path", type=str, required=True,
                        help="Output path for the training metrics file.")
    
    args = parser.parse_args()
    
    train_model_main(
        processed_data_path=args.processed_data_path,
        trained_model_path=args.trained_model_path,
        training_metrics_path=args.training_metrics_path
    )
