requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
docstring_parser==0.16
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.99.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-pipeline-components==2.20.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.21.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kfp==2.13.0
kfp-pipeline-spec==0.6.0
kfp-server-api==2.4.0
kubernetes==30.1.0
MarkupSafe==3.0.2
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
